Crate fxa_email_service

These are the developer docs for the Firefox Accounts email-sending service. For higher-level documentation, see the readme.

The project is compiled as a library that is linked against by two separate binaries:

fxa_email_send runs a Rocket server exposing an endpoint that enables callers to send email.

fxa_email_queues runs a process that loops infinitely, polling SQS queues for SES bounce, complaint and delivery notifications.

Modules

api
Parent scope for modules pertaining to the API layer.

db
Parent scope for modules pertaining to the database layer.

logging
Mozlog-compatible logging.

providers
Generic abstraction of specific email providers.

queues
Queue-processing abstractions.

settings
Application settings.

types
Parent scope for modules that implement miscellaneous generally-used types.