Enum fxa_email_service::app_errors::AppErrorKind

pub enum AppErrorKind {
    BadRequest,
    NotFound,
    MethodNotAllowed,
    UnprocessableEntity,
    TooManyRequests,
    InternalServerError,
    RocketError(Error),
    InvalidEmailParams,
    MissingEmailParams(String),
    InvalidProvider(String),
    ProviderError {
        name: String,
        description: String,
    },
    EmailParsingError(String),
    ComplaintError {
        address: EmailAddress,
        time: u64,
        problem: DeliveryProblem,
    },
    BounceSoftError {
        address: EmailAddress,
        time: u64,
        problem: DeliveryProblem,
    },
    BounceHardError {
        address: EmailAddress,
        time: u64,
        problem: DeliveryProblem,
    },
    AuthDbError(String),
    QueueError(String),
    InvalidNotificationType,
    MissingNotificationPayload(String),
    MissingSqsMessageFields {
        queue: String,
        field: String,
    },
    SqsMessageHashMismatch {
        queue: String,
        hash: String,
        body: String,
    },
    SqsMessageParsingError {
        queue: String,
        message: String,
        body: String,
    },
    DurationError(String),
    DbError(String),
    NotImplemented,
    HmacError(String),
    JsonError(String),
}

The specific kind of error that can occur.

Variants
BadRequest

400 Bad Request

NotFound

404 Not Found

MethodNotAllowed

405 Method Not Allowed

UnprocessableEntity

422 Unprocessable Entity

TooManyRequests

429 Too Many Requests,

InternalServerError

500 Internal Server Error

RocketError(Error)InvalidEmailParams

An error for invalid email params in the /send handler.

MissingEmailParams(String)

An error for missing email params in the /send handler.

InvalidProvider(String)

An error for invalid provider names.

ProviderError

An error for when we get an error from a provider.

Fields of ProviderError

name: String

description: String

EmailParsingError(String)

An error for when we have trouble parsing the email message.

ComplaintError

An error for when a bounce violation happens.

Fields of ComplaintError

address: EmailAddress

time: u64

problem: DeliveryProblem

BounceSoftError
Fields of BounceSoftError

address: EmailAddress

time: u64

problem: DeliveryProblem

BounceHardError
Fields of BounceHardError

address: EmailAddress

time: u64

problem: DeliveryProblem

AuthDbError(String)

An error occurred inside an auth db method.

QueueError(String)

An error for when an error happens on the queues process.

InvalidNotificationType

An error for when we get an invalid notification type in the queues process.

MissingNotificationPayload(String)

An error for when we get notification without a payload in the queues process.

MissingSqsMessageFields

An error for when we get SQS messages with missing fields.

Fields of MissingSqsMessageFields

queue: String

field: String

SqsMessageHashMismatch

An error for when the SQS message body does not match MD5 hash.

Fields of SqsMessageHashMismatch

queue: String

hash: String

body: String

SqsMessageParsingError

An error for when we can't parse the SQS message.

Fields of SqsMessageParsingError

queue: String

message: String

body: String

DurationError(String)

An error for when we get and invalid duration string.

DbError(String)

An error occured inside a db method.

NotImplemented

An error for when we try to access functionality that is not implemented.

HmacError(String)

An error occured while hashing a value.

JsonError(String)

An error occured while serializing or deserializing JSON.

Methods

impl AppErrorKind

pub fn http_status(&self) -> Status

Return a rocket response Status to be rendered for an error

pub fn errno(&self) -> Option<u16>

pub fn additional_fields(&self) -> Map<String, Value>