Module fxa_email_service::settings

Application settings.

Structs

AuthDb
Settings related to fxa-auth-db-mysql, which is used to store bounce, complaint and delivery notifications.

Aws
Settings for AWS.

AwsAccess
AWS access key type.

AwsKeys
AWS keys. These are sensitive data and will not be logged.

AwsRegion
AWS region type.

AwsSecret
AWS secret key type.

BaseUri
Base URI type.

DeliveryProblemLimit
A definition object for a bounce/complaint limit.

DeliveryProblemLimits
Controls the thresholds and behaviour for bounce and complaint reports.

Host
Host name or IP address type.

Log
Settings for logging.

Provider
Email provider settings.

Redis
Settings for Redis.

Sender
Controls the name and email address that are used for the From and Sender email headers.

SenderName
Sender name type.

Sendgrid
Settings for Sendgrid.

SendgridApiKey
Sendgrid API key type.

Sentry
Settings for Sentry.

SentryDsn
Sentry DSN type.

Settings
The root settings object.

Smtp
Settings for SMTP custom provider.

SmtpCredentials
Settings for SMTP custom provider optional credentials.

SocketLabs
Settings for SocketLabs.

SqsUrl
AWS SQS queue URL type.

SqsUrls
URLs for SQS queues.