Struct fxa_email_service::settings::Settings

pub struct Settings {
    pub authdb: AuthDb,
    pub aws: Aws,
    pub bouncelimits: BounceLimits,
    pub env: Env,
    pub hmackey: String,
    pub host: Host,
    pub log: Log,
    pub port: u16,
    pub provider: Provider,
    pub redis: Redis,
    pub secretkey: String,
    pub sender: Sender,
    pub sendgrid: Option<Sendgrid>,
    pub sentry: Option<Sentry>,
    pub smtp: Smtp,
    pub socketlabs: Option<SocketLabs>,
}

The root settings object.

Fields
authdb: AuthDb

Settings related to fxa-auth-db-mysql, which is used to store bounce, complaint and delivery notifications.

aws: Aws

Settings for AWS, including region, access keys and URLs for SQS queues.

bouncelimits: BounceLimits

Controls the thresholds and behaviour for bounce and complaint reports. If bounce limits are enabled, emails sent to offending addresses will fail with a 429 error.

env: Env

The env sets which environment we are in. It defaults to dev if not set.

hmackey: String

The HMAC key to use internally for hashing message ids. This is sensitive data and will not be logged.

host: Host log: Log

The logging settings, about level and formatting.

port: u16

The port this application is listening to.

provider: Provider

Settings controlling the default email provider.

redis: Redis

Settings for Redis, which is used to store metadata associated with a message.

secretkey: String

Setting for a secret key used by Rocket.

sender: Sender

Controls the name and email address that are used for the From and Sender email headers.

sendgrid: Option<Sendgrid>

Settings for Sendgrid.

sentry: Option<Sentry>

Settings for Sentry.

smtp: Smtp

Settings for SMTP custom provider.

socketlabs: Option<SocketLabs>

Settings for SocketLabs.

Methods

impl Settings

pub fn new() -> Result<Self, ConfigError>

Construct a Settings instance, populating it with data from the file system and local environment.

Precedence (earlier items override later ones):

Environment variables: $FXA_EMAIL_<UPPERCASE_KEY_NAME>

File: config/local.json

File: config/<$FXA_EMAIL_ENV>.json

File: config/default.json

Immediately before returning, the parsed config object will be logged to the console.

pub fn build_rocket_config(&self) -> Result<RocketConfig, RocketConfigError>

Create rocket configuration based on the environment variable. Defaults to dev mode if FXA_EMAIL_ENV is not set.