Module fxa_email_service::types::headers

Custom email header types.

On the one hand, this is an unfortunate side-effect of not being able to construct custom email headers using just methods from the emailmessage crate. On the other hand, we had a long term plan to whitelist permissable headers anyway, so this could ultimately form the basis for that logic too.

Structs

DeviceId

EmailSender

EmailService

FlowBeginTime

FlowId

Link

RecoveryCode

ReportSigninLink

ServiceId

SesConfigurationSet

SesMessageTags

SigninVerifyCode

TemplateName

TemplateVersion

Uid

UnblockCode

VerifyCode