Module fxa_email_service::types::validate

Common validation logic.

All functions are predicates, returning true if a value is valid and false if it isn't. Note that the intention is not to provide perfect validation in each case. Mostly it is to rule out obvious mistakes when setting values in config or wiring in request parameters.

Functions

aws_access
Validate an AWS access key.

aws_region
Validate an AWS region.

aws_secret
Validate an AWS secret key.

base_uri
Validate a base URI.

email_address
Validate an email address.

host
Validate a host name or IP address.

logging_format
Validate logging format.

logging_level
Validate logging level.

sender_name
Validate a sender name.

sendgrid_api_key
Validate a Sendgrid API key.

sentry_dsn
Validate a Sentry DSN.

sqs_url
Validate an AWS SQS URL.