Enum fxa_email_service::app_errors::AppErrorKind

pub enum AppErrorKind {
    BadRequest,
    NotFound,
    MethodNotAllowed,
    UnprocessableEntity,
    TooManyRequests,
    InternalServerError,
    RocketError(Error),
    InvalidEmailParams,
    MissingEmailParams(String),
    InvalidProvider(String),
    ProviderError {
        name: String,
        description: String,
    },
    EmailParsingError(String),
    BounceComplaintError {
        address: String,
        bounce: Option<BounceRecord>,
    },
    BounceSoftError {
        address: String,
        bounce: Option<BounceRecord>,
    },
    BounceHardError {
        address: String,
        bounce: Option<BounceRecord>,
    },
    DbError(String),
    NotImplemented,
}

The specific kind of error that can occur.

Variants
BadRequest

400 Bad Request

NotFound

404 Not Found

MethodNotAllowed

405 Method Not Allowed

UnprocessableEntity

422 Unprocessable Entity

TooManyRequests

429 Too Many Requests,

InternalServerError

500 Internal Server Error

RocketError(Error)InvalidEmailParams

An error for invalid email params in the /send handler.

MissingEmailParams(String)

An error for missing email params in the /send handler.

InvalidProvider(String)

An error for invalid provider names.

ProviderError

An error for when we get an error from a provider.

Fields of ProviderError

name: String

description: String

EmailParsingError(String)

An error for when we have trouble parsing the email message.

BounceComplaintError

An error for when a bounce violation happens.

Fields of BounceComplaintError

address: String

bounce: Option<BounceRecord>

BounceSoftError
Fields of BounceSoftError

address: String

bounce: Option<BounceRecord>

BounceHardError
Fields of BounceHardError

address: String

bounce: Option<BounceRecord>

DbError(String)

An error for when an error happens on a request to the db.

NotImplemented

An error for when we try to access functionality that is not implemented.

Methods

impl AppErrorKind

pub fn http_status(&self) -> Status

Return a rocket response Status to be rendered for an error

pub fn errno(&self) -> Option<i32>

pub fn additional_fields(&self) -> Map<String, Value>