Struct fxa_email_service::settings::Settings

pub struct Settings {
    pub authdb: AuthDb,
    pub aws: Aws,
    pub bouncelimits: BounceLimits,
    pub hmackey: String,
    pub logging: Logging,
    pub provider: Provider,
    pub redis: Redis,
    pub sender: Sender,
    pub sendgrid: Option<Sendgrid>,
    pub smtp: Smtp,
}

The root settings object.

Fields
authdb: AuthDb

Settings related to fxa-auth-db-mysql, which is used to store bounce, complaint and delivery notifications.

aws: Aws

Settings for AWS, including region, access keys and URLs for SQS queues.

bouncelimits: BounceLimits

Controls the thresholds and behaviour for bounce and complaint reports. If bounce limits are enabled, emails sent to offending addresses will fail with a 429 error.

hmackey: String

The HMAC key to use internally for hashing message ids. This is sensitive data and will not be logged.

logging: Logging

The logging format to use, can be "mozlog", "pretty" or "null".

provider: Provider

The default email provider to use, can be "ses", "sendgrid" or "mock". Note that this setting can be overridden on a per-request basis.

redis: Redis

Settings for Redis, which is used to store metadata associated with a message.

sender: Sender

Controls the name and email address that are used for the From and Sender email headers.

sendgrid: Option<Sendgrid>

Settings for Sendgrid.

smtp: Smtp

Settings for SMTP custom provider.

Methods

impl Settings

pub fn new() -> Result<Self, ConfigError>

Construct a Settings instance, populating it with data from the file system and local environment.

Precedence (earlier items override later ones):

Environment variables: $FXA_EMAIL_<UPPERCASE_KEY_NAME>

File: config/local.json

File: config/<$NODE_ENV>.json

File: config/default.json

$NODE_ENV is used so that this service automatically picks up the appropriate state from our existing node.js ecosystem, without needing to manage an extra environment variable.

Immediately before returning, the parsed config object will be logged to the console.